refactor: Rename block "bootstrap" to "import blocks"
Due to overloading of the term "bootstrap" the importation of blocks
from a provided directory no longer uses this term.

This closes #612.
