+++ docs/src/user-guides/bootstrapping-state.md
@@ -2,10 +2,10 @@
 
 Instead of starting the state from the genesis block and requesting all blocks from peers, the `neptune-core` node can bootstrap its state from a directory containing raw block data.
 
-This can be done by using the `--bootstrap-from-directory=<DIR>` CLI argument when launching `neptune-core`. The directory must contain raw block data of the form generated by the node during normal operations. The file names should have the form `blk0.dat`, `blk1.dat` etc.
+This can be done by using the `--import-blocks-from-directory=<DIR>` CLI argument when launching `neptune-core`. The directory must contain raw block data of the form generated by the node during normal operations. The file names should have the form `blk0.dat`, `blk1.dat` etc.
 
-Once the node has read all blocks this way, the directory containing the block data used for bootstrapping can be deleted, as the process copies all block data.
+Once the node has read all blocks this way, the directory containing the block data used for block importing can be deleted, as the process copies all block data.
 
-By default, the blocks read this way will be validated. Skipping this validation step can speedup initial synchronization significantly. To skip the validation step, add the argument `--disable-bootstrap-block-validation`. If you skip the validation step, this process can no longer be considered trustless, and `neptune-core` may end up in an unrecoverable state if the source data is somehow corrupted.
+By default, the blocks read this way will be validated. Skipping this validation step can speedup initial synchronization significantly. To skip the validation step, add the argument `--disable-validation-in-block-import`. If you skip the validation step, this process can no longer be considered trustless, and `neptune-core` may end up in an unrecoverable state if the source data is somehow corrupted.
 
 A torrent of the latest snapshot can always be downloaded from <http://neptunefundamentals.org:42580/latest-snapshot.torrent>. This snapshot is updated weekly.
